Function Reference

CARS.BlackBoxOptions — Type

Base.@kwdef mutable struct BlackBoxOptions
    MaxFuncEvals::Int = 1000                    # Stop optimizing after reaching this number of iterations
    TargetFitness::Float64 = 1e-7                 # Stop optimizing after reachting this Fitness level
    # FitnessTolerance not yet implemented
    FitnessTolerance::Float64 = 1e-7            # If StoppingCriterion==:tol, stop if change between iterations is lower than this value
    TraceMode::Symbol = :silent                 # Output of blackboxoptim during optimization, see docs of BlackBoxOptim.jl for choices
    Method::Symbol = :adaptive_de_rand_1_bin    # Method used for optimization, see docs of BlackBoxOptim.jl for choices
    StoppingCriterion::Symbol = :best           # If set to :best, iteration will stop if BestFitness is reached. If set to :tol, it will stop if difference between iterations is smaller than FitnessTolerance
    PopulationSize::Int = 5                    # Population Size for methods which actually use a population. No effect for others.
end

Specify options for blackBoxFit

CARS.CARSOptions — Type

mutable struct CARSOptions

Specifies global options that control certain aspects of CARS.jl

fittype::Symbol: either :sqrt to take the square root of the signal intensity (after preprocessing) or anything else (use the signal directly). Affects preprocess_spectra, calcResidual, generateSpectrum! and noiseModel!.

γf::AbstractFloat: Controls rediscretization after convolution with the compression kernel in generateLibrary and simplifyLibrary!. Higher values means higher resolution of wavenumber axis. Defaults to 5.

interp: Control interpolation scheme for some of the operations. Stick with Linear().

conv_margin: Specifies a safety margin for the convolution with the apparatus function to avoid edge effects after simplifying a library. Applied as a factor to the instrumental linewidth.

CARS.FitParams — Type

Base.@kwdef mutable struct FitParams
    T                          # Temperature as value
    X::Dict{String,Any}        # Dictionary of Mole Fractions of resonant species
    chiNR=0                      # Non-resonant signal of resonant species. No need to set this manually, will be filled with correct values automatically
    chiB=8.5                     # Non-resonant susceptibility of the buffer gas    
    AdditionalFrequencyOffset=0  # For DP CARS Spectra. Additional shift (to what's already in the library) of Pump region 2 relative to pump region 1
    LineWidth::Dict{String,Any} # Dict of Laser linewidth(s), "P1" for pump region 1, "P2" for pump region 2
    Instrumental=0             # Instrumental LineWidth of spectrometer
    WavenumberPolynomial::Vector{Any}=[2200.,0.5,0.]     # Vector of polynomial parameter to fit the wavenumber axis, Syntax [a, b, c] will be interpreted as a+x^1*b+x^2*c
    WavenumberShift=0          # Shift of calibrated wavenumber axis (zero order)
    Spectra::SpectraStruct=SpectraStruct([0.],[0.],[0.],[0.],0,"unknown")     # After fitting, this will contain the original experimental spectrum (after area normalization) along with the fit
end

Struct containing all relevant degrees of freedom for simulatin/fitting spectra. Every parameter in here (with the exception of Spectra) can be used in the fit.

CARS.IpoptOptions — Type

Base.@kwdef mutable struct IpoptOptions
    max_iter::Int = 1000                        # Abort after a certain number of iterations
    max_wall_time::Float64 = 120.               # Abort after reaching a certain wall time
    tol::Float64 = 1e-8                         # Abort when reaching this tolerance (see Ipopt Documentation, 1e-7 is the default value)
    print_level::Int = 3                        # Amount of information during optimization.
    mu_strategy::String = "adaptive"            # "monotone" or "adaptive".
end

Specify options for gradientFit Some options passed to the Ipopt Optimizer. If you wish to extend this, check out the official Ipopt documentation Values in this struct are all passed to the optimizer by name.

CARS.Library — Type

mutable struct Library
    species::Vector{String}  # name of resonant species in this library
    ROI::Vector                 # Region of Interest (excluding extension)
    AdditionalFrequencyOffset     # FrequencyOffset relative to PumpRegion 1
    T::Vector           # Vector of temperatures
    P::Float32          # Pressure
    ω::Vector{Float32}  # Wavenumberarray after compression
    ωres::Float32       # Wavenumber resolution used in library
    γc::Float32          # width of preconvolution kernel
    χC::Matrix{Array{Float32}} # compressed complex cross-terms
    χM::Array{Float32} # Compressed magnitude
    χI::Array{Float32} # Compressed imaginary components
    χR::Array{Float32} # Compressed real components
    χNR::Vector{Float32}  # (Purely real) non-resonant susceptibility of species in library. Already converted in cm3/molecule
end

Struct containing all relevant data for a single pumping region.

CARS.PreProcSpectrum — Type

mutable struct PreProcSpectrum
    I::AbstractVector{Float64}          # area normalized intensity
    w::AbstractArray{Float64}           # weight used for fit
    A::Float64                          # original area (used to scale the variance for noise weighting)
    ErrorCode::Int8                     # specify errorcode after preprocessing (0: no error, 1:saturated, 2: signal too low, 3: breakdown)
end

CARS.RawSpectra — Type

mutable struct RawSpectra
    I::AbstractArray{Float64}      # signal
    BG::AbstractArray{Float64}     # background
    SingleShotNR::AbstractArray{Float64}     # preparation for single shot NR correction. not yet implemented.
end

Struct containing raw spectra after loading from a file.

CARS.argonModel! — Method

argonModel!(preprocessed,nr,roi)

Extract a weight based on Stokes noise estimated from measurements in Argon.

CARS.benchmarkGradient — Method

benchmarkGradient(IE::PreProcSpectrum,startingsolution::FitParams,lower::FitParams,upper::FitParams,libs::Library...)

Function to compute the gradient of the residual with the same argument structure as the fitting operations. Use e.g with @benchmark or simply @time. Helpful when comparing an optimized handling of an operation on duals with a working reference.

CARS.benchmarkValue — Method

benchmarkValue(IE::PreProcSpectrum,startingsolution::FitParams,lower::FitParams,upper::FitParams,libs::Library...)

Function to compute the value of the residual with the same argument structure as the fitting operations. Use e.g with @benchmark or simply @time. Helpful when comparing an optimized handling of an operation on duals with a working reference.

CARS.blackBoxFit — Method

blackBoxFit(IE::PreProcSpectrum,startingsolution::FitParams,lower::FitParams,upper::FitParams,libs::Library...;options::BlackBoxOptions=BlackBoxOptions(),showplots=false)

Use BlackBoxOptim.jl for fitting.

CARS.calcResidual — Method

calcResidual(spec::PreProcSpectrum,libs,params::FitParams,loc,x::T...) where {T<:Real}

Calculates the resiudal between preprocessed experimental spectrum spec and a simulated spectrum defined by params. Requires loc to interpret the vector x as fitting parameters to store in params struct.

Note

All methods wihtin this function have to be compatible with T<:Real to be compatible with FowardDiff.Duals!

CARS.combinedFit — Method

combinedFit(IE::PreProcSpectrum,startingsolution::FitParams,lower::FitParams,upper::FitParams,libs::Library...;blackboxopt::BlackBoxOptions=BlackBoxOptions(MaxFuncEvals=200,TraceMode=:silent),ipoptopt::IpoptOptions=IpoptOptions(print_level=0),showplots=false) Use blackBoxFit first which provides a starting solution for gradientFit for final fitting.

CARS.fastLinearInt — Method

fastLinearInt(wlib::AbstractArray{R},data::AbstractArray{<:ForwardDiff.Dual{T,V,K}},wexp::AbstractArray{<:AbstractFloat};extrap=false)  where {R,T,V,K}

Do a fast linear interpolation of a Dual number array data from the original grid wlib to the destination grid wexp.

CARS.findNeighbours — Method

findNeighbours(wexp,wlib;extrap=false)

Finds neighbours of each element in wexp in the finer array wlib. Used for fastLinearInt.

CARS.generateLibrary — Method

generateLibrary(T,γc,getSuscs...)

Generates libraries in the temperature range T, e.g. T=280:20:2000 using the compression kernel γc and the supplied methods getSuscs...

CARS.generateSpectrum! — Method

generateSpectrum!(params::FitParams,libs...;takesqrt=CARSopt.fittype)

Convenience function to simulate spectra outside fitting mode.

CARS.generateSpectrum! — Method

generateSpectrum!(params::FitParams,a::T,loc,libs...;takesqrt=CARSopt.fittype) where {T<:Real}

Generates a spectrum based on the parameters provided by params using the libraries libs.... The loc array is required to identify if some of the degrees of freedom are required, e.g. frequency offset, where the interpolation is not performed if not required.

Note

All methods wihtin this function have to be compatible with T<:Real to be compatible with FowardDiff.Duals!

CARS.gradientFit — Method

gradientFit(IE::PreProcSpectrum,startingsolution::FitParams,lower::FitParams,upper::FitParams,libs::Library...;options::IpoptOptions=IpoptOptions(),showplots=false)

Use JuMP+Ipopt for fitting.

CARS.isFitParam — Method

isFitParam(loc,name)

Returns true if name is a fit parameter by detecting if it is in the loc array created by vectorizeParams!

CARS.noiseModel! — Method

noiseModel!(type::Symbol,x...)

Abstraction to extract a noiseModel with various methods. Only implemented at the moment: :argon which calls argonModel!

CARS.preprocess_spectra — Method

preprocess_spectra(signal::RawSpectra,nr_signal::RawSpectra,roi;errorCriteria::ErrorCriteria=ErrorCriteria())

Preprocesses spectra passed to this function as signal. It uses the some criteria to exclude spectra from fitting, e.g. when the signal is too high or too low or if a breakdown occured. Check out the example fit_experimental_data.jl for how to adapt the ErrorCriteria to your needs. Signals are normalized to the supplied non-resonant signal nr_signal of type RawSpectra and, if specified in CARSOptions the square root is taken.

CARS.save — Method

This function can be used to conveniently store data in either a .MAT or .JLD2 file, depending on what you would like to use for further processing If the file exists, the data is added to the file.

CARS.simplifyLibrary! — Method

This function uses the linewidth and additionalfrequencyoffset in fitparams and applies them to the library. This allows to treat these as fixed parameters for the fit and make it much(!) faster

CARS.vectorizeParams! — Method

vectorizeParams!